Remove the lens dependency from vinyl.
The only feature lost is the definition of rIso that defines an Iso'
from the lens package.

Replaced mtl dependency with transformers. The only piece we use is
Data.Functor.Identity, which should end up in base soon.
