+++ Data/Vinyl.hs
@@ -5,13 +5,11 @@ module Data.Vinyl
   , module Data.Vinyl.Rec
   , module Data.Vinyl.Relation
   , module Data.Vinyl.Classes
-  , module Data.Vinyl.MiniLens
   ) where
 
 import           Data.Vinyl.Classes
 import           Data.Vinyl.Field
 import           Data.Vinyl.Lens
-import           Data.Vinyl.MiniLens
 import           Data.Vinyl.Rec
 import           Data.Vinyl.Relation
 import           Data.Vinyl.Witnesses
